RSS_Feed_MCP/requirements.txt
pyyaml
ascii_colors
fastmcp
feedparser
pipmaster